[PriceRules] add priority to product-price-rules
