+++ src/CoreShop/Bundle/ProductBundle/Form/Type/ProductPriceRuleType.php
@@ -15,6 +15,7 @@ namespace CoreShop\Bundle\ProductBundle\Form\Type;
 use CoreShop\Bundle\ResourceBundle\Form\Type\ResourceTranslationsType;
 use CoreShop\Bundle\RuleBundle\Form\Type\RuleType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 
@@ -30,6 +31,7 @@ final class ProductPriceRuleType extends RuleType
                 'entry_type' => ProductPriceRuleTranslationType::class,
             ])
             ->add('name', TextareaType::class)
+            ->add('priority', IntegerType::class)
             ->add('active', CheckboxType::class)
             ->add('conditions', ProductPriceRuleConditionCollectionType::class)
             ->add('actions', ProductPriceRuleActionCollectionType::class);
