+++ src/CoreShop/Bundle/ProductBundle/Form/Type/ProductSpecificPriceRuleType.php
@@ -15,7 +15,7 @@ namespace CoreShop\Bundle\ProductBundle\Form\Type;
 use CoreShop\Bundle\ResourceBundle\Form\Type\ResourceTranslationsType;
 use CoreShop\Bundle\RuleBundle\Form\Type\RuleType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
-use Symfony\Component\Form\Extension\Core\Type\NumberType;
+use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 
@@ -33,7 +33,7 @@ final class ProductSpecificPriceRuleType extends RuleType
             ->add('name', TextareaType::class)
             ->add('inherit', CheckboxType::class)
             ->add('active', CheckboxType::class)
-            ->add('priority', NumberType::class)
+            ->add('priority', IntegerType::class)
             ->add('conditions', ProductSpecificPriceRuleConditionCollectionType::class)
             ->add('actions', ProductSpecificPriceRuleActionCollectionType::class);
     }
