+++ src/CoreShop/Bundle/ProductBundle/Resources/config/serializer/Model.AbstractPriceRule.yml
@@ -1,0 +1,33 @@
+CoreShop\Component\Product\Model\AbstractPriceRule:
+    exclusion_policy: ALL
+    xml_root_name: price_rule
+    properties:
+        id:
+            expose: true
+            type: integer
+            xml_attribute: true
+            groups: [List, Detailed]
+        name:
+            expose: true
+            type: string
+            groups: [List, Detailed]
+        active:
+            expose: true
+            type: boolean
+            groups: [Detailed]
+        priority:
+            expose: true
+            type: integer
+            groups: [Detailed]
+        conditions:
+            expose: true
+            type: array
+            groups: [Detailed]
+        actions:
+            expose: true
+            type: array
+            groups: [Detailed]
+        translations:
+            expose: true
+            type: array
+            groups: [Detailed]
