+++ src/CoreShop/Bundle/ProductBundle/Resources/config/serializer/Model.PriceRuleTranslation.yml
@@ -1,0 +1,8 @@
+CoreShop\Component\Product\Model\PriceRuleTranslation:
+    exclusion_policy: ALL
+    xml_root_name: price_rule_translation
+    properties:
+        label:
+            expose: true
+            type: string
+            groups: [Detailed]
