+++ src/CoreShop/Bundle/ProductBundle/Resources/config/serializer/Model.ProductPriceRule.yml
@@ -1,29 +1,3 @@
 CoreShop\Component\Product\Model\ProductPriceRule:
     exclusion_policy: ALL
     xml_root_name: product_price_rule
-    properties:
-        id:
-            expose: true
-            type: integer
-            xml_attribute: true
-            groups: [List, Detailed]
-        name:
-            expose: true
-            type: string
-            groups: [List, Detailed]
-        active:
-            expose: true
-            type: boolean
-            groups: [List, Detailed]
-        conditions:
-            expose: true
-            type: array
-            groups: [Detailed]
-        actions:
-            expose: true
-            type: array
-            groups: [Detailed]
-        translations:
-            expose: true
-            type: array
-            groups: [Detailed]
