+++ src/CoreShop/Bundle/ProductBundle/Resources/config/serializer/Model.ProductPriceRuleTranslation.yml
@@ -1,8 +1,3 @@
 CoreShop\Component\Product\Model\ProductPriceRuleTranslation:
     exclusion_policy: ALL
     xml_root_name: product_price_rule_translation
-    properties:
-        label:
-            expose: true
-            type: string
-            groups: [Detailed]
