+++ src/CoreShop/Component/Product/Model/ProductSpecificPriceRule.php
@@ -15,19 +15,8 @@ namespace CoreShop\Component\Product\Model;
 use CoreShop\Component\Resource\Model\TranslatableTrait;
 use CoreShop\Component\Rule\Model\RuleTrait;
 
-class ProductSpecificPriceRule implements ProductSpecificPriceRuleInterface
+class ProductSpecificPriceRule extends AbstractPriceRule implements ProductSpecificPriceRuleInterface
 {
-    use RuleTrait;
-    use TranslatableTrait {
-        __construct as private initializeTranslationsCollection;
-        getTranslation as private doGetTranslation;
-    }
-
-    /**
-     * @var int
-     */
-    protected $id;
-
     /**
      * @var int
      */
@@ -37,24 +26,6 @@ class ProductSpecificPriceRule implements ProductSpecificPriceRuleInterface
      * @var bool
      */
     protected $inherit = false;
-
-    /**
-     * @var int
-     */
-    protected $priority = 0;
-
-    public function __construct()
-    {
-        $this->initializeTranslationsCollection();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
 
     /**
      * {@inheritdoc}
@@ -90,54 +61,6 @@ class ProductSpecificPriceRule implements ProductSpecificPriceRuleInterface
         $this->inherit = $inherit;
 
         return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getPriority()
-    {
-        return $this->priority;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setPriority($priority)
-    {
-        $this->priority = $priority;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLabel($language = null)
-    {
-        return $this->getTranslation($language)->getLabel();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLabel($label, $language = null)
-    {
-        $this->getTranslation($language)->setLabel($label);
-    }
-
-    /**
-     * @param null $locale
-     * @param bool $useFallbackTranslation
-     *
-     * @return ProductSpecificPriceRuleTranslationInterface
-     */
-    public function getTranslation($locale = null, $useFallbackTranslation = true)
-    {
-        /** @var ProductSpecificPriceRuleTranslationInterface $translation */
-        $translation = $this->doGetTranslation($locale, $useFallbackTranslation);
-
-        return $translation;
     }
 
     /**
