+++ src/CoreShop/Component/Product/Repository/ProductSpecificPriceRuleRepositoryInterface.php
@@ -14,9 +14,8 @@ namespace CoreShop\Component\Product\Repository;
 
 use CoreShop\Component\Product\Model\ProductInterface;
 use CoreShop\Component\Product\Model\ProductSpecificPriceRuleInterface;
-use CoreShop\Component\Rule\Repository\RuleRepositoryInterface;
 
-interface ProductSpecificPriceRuleRepositoryInterface extends RuleRepositoryInterface
+interface ProductSpecificPriceRuleRepositoryInterface extends PriceRuleRepositoryInterface
 {
     /**
      * @param ProductInterface $product
